src-dummy/pyproject.toml
[project]
name = "tcpbroker"
version = "0.1.0"
description = "TCP as Broker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
]
